[CHANGED] Model tests separated into their own files
